LayoutTests/fast/scrolling/mac/momentum-event-sequence-expected.txt

PASS window.scrollY is 56
PASS successfullyParsed is true

TEST COMPLETE
